// locadora-app/src/app/models/Locadora/locadora.create.model.ts
// import { Endereco } from './endereco.model';
export interface CreateLocadora {
    nomeFantasia: string;
    razaoSocial: string;
    cnpj: string;
    email: string;
    telefone: string;    
    cep: string;
    rua: string;
    numero: string;
    bairro: string;
    estado: string;
    cidade: string;
}